python-mysql/requirements.txt
mysql-connector-python==8.0.26
protobuf==3.18.0
